Added beginning of project detail dashboard.
The site has been configured to accept project details at this URL: http://localhost/projects/1/my-project Two static line charts have been added to the dashboard. The styles which are used on the home page and the details page have been consolidated into Site.css, with the most specific styles going to Home.css and Project.css. The banners which are used on both pages are now located inside of the Images/banners folder.