Finished project dashboard design, added breadcrumb
The project dashboard design has been completed. The temporary charting components are in place and all of the information which is intended to be displayed has been implemented as static data into the dashboard. The _Home and _Project views now support breadcrumbs as a form of navigation near the site logo.