Uses flat-file for field ID database.
This project is no longer dependent on the Entity Framework to keep track of a very small database. This database has been replaced with a simple, flat-file database. All references and files related to the old Entity-driven database has been removed. A new factory class was added to manage this flat-file database. The TasksFactory class has been updated to include the amount of time worked and expected to work on a task along with the task completion percentage whenever supplying the application with data from its model.